+++ 00_main.tf
@@ -4,20 +4,6 @@ provider "aws" {
   profile                 = "personal"
 }
 
-data "aws_caller_identity" "current" {}
-
-output "account_id" {
-  value = "${data.aws_caller_identity.current.account_id}"
-}
-
-output "caller_arn" {
-  value = "${data.aws_caller_identity.current.arn}"
-}
-
-output "caller_user" {
-  value = "${data.aws_caller_identity.current.user_id}"
-}
-
 terraform {
   backend "s3" {
     shared_credentials_file = "~/.aws/credentials"