+++ 99_variables.tf
@@ -18,24 +18,36 @@ variable "iot_topic" {
   description = "IoT topic that would be used to publish messages"
 }
 
+### Lambda ###
+variable "lambda_index_to_es_role_name" {
+  description = "Name of the role that will be attached to the Lambda function that indexes to Elasticsearch"
+}
+variable "lambda_index_to_es_policy_name" {
+  description = "Name of the policy that will be attached to the role for the Lambda function that indexes to Elasticsearch"
+}
+
 ### Elasticsearch ###
-variable "domain_name" {
+variable "es_domain_name" {
   description = "Elasticsearch service domain name"
 }
-variable "elasticsearch_version" {
+variable "es_elasticsearch_version" {
   description = "Elasticsearch version"
 }
-variable "instance_type" {
+variable "es_instance_type" {
   description = "Elasticsearch service instance type"
 }
-variable "ebs_enabled" {
+variable "es_ebs_enabled" {
   description = "Enable or disable ebs for Elasticsearch service"
 }
-variable "volume_size" {
+variable "es_volume_size" {
   description = "Ebs size for Elasticsearch service"
 }
 
-## Tags
+variable "es_tls_security_policy" {
+  description = "TLS Security policy for Elasticsearch"
+}
+
+### Tags ###
 variable "tags" {
   type = map
 }