+++ scripts/save_outputs_terraform.py
@@ -1,5 +1,14 @@
 import subprocess
 import json
+import os
+
+if not os.path.isdir("../certs"):
+    os.mkdir("../certs", mode=0o755)
+    os.chmod("../certs", mode=0o755)
+
+if not os.path.isdir("../config"):
+    os.mkdir("../config", mode=0o755)
+    os.chmod("../config", mode=0o755)
 
 p = subprocess.check_output(["terraform", "output", "-json"], cwd="../.")
 json = json.loads(p.decode("utf-8"))  # print(type(p.decode("utf-8")))
@@ -7,6 +16,7 @@ json = json.loads(p.decode("utf-8"))  # print(type(p.decode("utf-8")))
 cert = json["iot_certificate_pem"]["value"]
 private_key = json["iot_certificate_private_key"]["value"]
 public_key = json["iot_certificate_public_key"]["value"]
+elasticsearch_domain = json["elastisearch_domain_endpoint"]["value"]
 
 fcert = open('../certs/cert.crt', 'w')
 fcert.write(cert)
@@ -19,3 +29,7 @@ fpriv.close()
 fpub = open('../certs/public_key.pem', 'w')
 fpub.write(public_key)
 fpub.close()
+
+fes = open('../config/elasticsearch_domain', 'w')
+fes.write(elasticsearch_domain)
+fes.close()